implement createExpense() method and tested,
