--- src/main/java/com/beeWallet/beeWallet/enums/ExpenseEnum.java
@@ -1,9 +1,0 @@
-package com.beeWallet.beeWallet.enums;
-
-public enum ExpenseEnum {
-    HIVE_AND_ELEMENTS,
-    BEE,
-    HONEY_EXTRACTOR,
-    SUPPLEMENTS,
-    FOOD_FOR_BEES
-}
