--- src/main/java/com/beeWallet/beeWallet/enums/IncomenEnum.java
@@ -1,9 +1,0 @@
-package com.beeWallet.beeWallet.enums;
-
-public enum IncomenEnum {
-    HONEY,
-    PROPOLIS,
-    BEE_POLLEN,
-    BEES,
-    HIVE_AND_ELEMENTS
-}
