+++ src/main/java/com/beeWallet/beeWallet/service/ExpenseService.java
@@ -1,6 +1,8 @@
 package com.beeWallet.beeWallet.service;
 
+import com.beeWallet.beeWallet.mapper.ExpenseMapper;
 import com.beeWallet.beeWallet.repository.ExpenseRepository;
+import com.beeWallet.beeWallet.repository.entity.ExpenseEntity;
 import com.beeWallet.beeWallet.web.model.ExpenseModel;
 import org.springframework.stereotype.Service;
 
@@ -12,16 +14,21 @@ public class ExpenseService {
     private static final Logger LOGGER = Logger.getLogger(ExpenseService.class.getName());
 
     private ExpenseRepository expenseRepository;
+    private ExpenseMapper expenseMapper;
 
-    public ExpenseService(ExpenseRepository expenseRepository) {
+    public ExpenseService(ExpenseRepository expenseRepository, ExpenseMapper expenseMapper) {
         this.expenseRepository = expenseRepository;
+        this.expenseMapper = expenseMapper;
     }
 
     // C - create
-    public void create(ExpenseModel expanse) {
+    public ExpenseModel create(ExpenseModel expanse) {
         LOGGER.info("create(" + expanse + ")");
-
-        LOGGER.info("create(...)");
+        ExpenseEntity expenseEntity = expenseMapper.from(expanse);
+        ExpenseEntity savedExpenseEntity = expenseRepository.save(expenseEntity);
+        ExpenseModel savedExpenseModel = expenseMapper.from(savedExpenseEntity);
+        LOGGER.info("create(...) " + savedExpenseModel);
+        return savedExpenseModel;
     }
 
     // R - read
